# static/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoOnCall",
    "description": "基于 LangChain 的智能oncall运维系统",
    "version": "1.2.1"
  },
  "paths": {
...
    "/api/approvals/pending": {
      "get": {
        "tags": [
          "AIOps人工审批"
        ],
        "summary": "List Pending Approvals",
        "description": "List the operator approval queue.",
        "operationId": "list_pending_approvals_api_approvals_pending_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Incident Id"
            }
          },
          {
            "name": "include_approved_actions",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Approved Actions"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{incident_id}/approval": {
      "get": {
        "tags": [
          "AIOps人工审批"
        ],
        "summary": "List Incident Approvals",
        "description": "List approval requests for one incident.",
        "operationId": "list_incident_approvals_api_incidents__incident_id__approval_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Incident Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "approved",
                    "rejected",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentApprovalListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AIOps人工审批"
        ],
        "summary": "Submit Incident Approval",
        "description": "Approve or reject the explicitly identified request for an incident.",
        "operationId": "submit_incident_approval_api_incidents__incident_id__approval_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Incident Id"
            }
          },
          {
            "name": "X-AutoOnCall-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Autooncall-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalDecisionRequest": {
        "properties": {
          "approval_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Approval Id"
          },
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "reject",
              "cancel"
            ],
            "title": "Decision"
          },
          "decided_by": {
            "type": "string",
            "maxLength": 120,
            "title": "Decided By",
            "default": "operator"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "approval_id",
          "decision"
        ],
        "title": "ApprovalDecisionRequest",
        "description": "API payload for approving or rejecting a pending action."
      },
      "ApprovalDecisionResponse": {
        "properties": {
          "approval": {
            "$ref": "#/components/schemas/ApprovalRequest"
          }
        },
        "type": "object",
        "required": [
          "approval"
        ],
        "title": "ApprovalDecisionResponse",
        "description": "Response after approving or rejecting one approval request."
      },
      "ApprovalListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApprovalRequest"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "ApprovalListResponse",
        "description": "List response for pending approval requests."
      },
      "ApprovalRequest": {
        "properties": {
          "approval_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Approval Id"
          },
          "incident_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Incident Id"
          },
          "action": {
            "type": "string",
            "maxLength": 1000,
            "title": "Action"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Level"
          },
          "reason": {
            "type": "string",
            "maxLength": 4000,
            "title": "Reason",
            "default": ""
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "cancelled"
            ],
            "title": "Status",
            "default": "pending"
          },
          "step_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Id"
          },
          "tool_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Name"
          },
          "risk_policy_version": {
            "type": "string",
            "maxLength": 64,
            "title": "Risk Policy Version",
            "default": ""
          },
          "change_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChangePlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "requested_by": {
            "type": "string",
            "maxLength": 120,
            "title": "Requested By",
            "default": "aiops-agent"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decided_by_principal_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By Principal Id"
          },
          "decision_reason": {
            "type": "string",
            "maxLength": 2000,
            "title": "Decision Reason",
            "default": ""
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "projection_pending": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Projection Pending"
          },
          "expires_in_seconds": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 1.0,
            "title": "Expires In Seconds",
            "default": 3600
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "action",
          "risk_level"
        ],
        "title": "ApprovalRequest",
        "description": "Human approval request for non-read-only or risky actions."
      },
...
      "ChangePlan": {
        "properties": {
          "change_plan_id": {
            "type": "string",
            "title": "Change Plan Id"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Level",
            "default": "medium"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "approved",
              "rejected",
              "cancelled"
            ],
            "title": "Status",
            "default": "draft"
          },
          "pre_checklist": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pre Checklist"
          },
          "execution_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Execution Steps"
          },
          "rollback_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rollback Steps"
          },
          "verification_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Verification Steps"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ChangeStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "rollback_plan": {
            "items": {
              "$ref": "#/components/schemas/ChangeStep"
            },
            "type": "array",
            "title": "Rollback Plan"
          },
          "remediation_playbook": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RemediationPlaybook"
              },
              {
                "type": "null"
              }
            ]
          },
          "observe_metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Observe Metrics"
          },
          "blast_radius": {
            "type": "string",
            "title": "Blast Radius",
            "default": ""
          },
          "expires_in_seconds": {
            "type": "integer",
            "title": "Expires In Seconds",
            "default": 3600
          },
          "manual_execution_required": {
            "type": "boolean",
            "title": "Manual Execution Required",
            "default": true
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": "Agent 只生成变更计划草案，不自动执行生产动作。"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "action"
        ],
        "title": "ChangePlan",
        "description": "A non-executing production change plan attached to approvals and reports."
      },
...
      "ChangeStep": {
        "properties": {
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type",
            "default": "manual"
          },
          "target": {
            "type": "string",
            "title": "Target",
            "default": ""
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name",
            "default": ""
          },
          "input_args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Args"
          },
          "expected_result": {
            "type": "string",
            "title": "Expected Result",
            "default": ""
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Level",
            "default": "medium"
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": true
          },
          "can_dry_run": {
            "type": "boolean",
            "title": "Can Dry Run",
            "default": true
          },
          "rollback_step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rollback Step Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "succeeded",
              "failed",
              "blocked",
              "skipped"
            ],
            "title": "Status",
            "default": "pending"
          }
        },
        "type": "object",
        "title": "ChangeStep",
        "description": "Structured execution or rollback step inside a human-approved change plan."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentApprovalListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApprovalRequest"
            },
            "type": "array",
            "title": "Items"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          }
        },
        "type": "object",
        "required": [
          "incident_id"
        ],
        "title": "IncidentApprovalListResponse",
        "description": "List response for approval requests scoped to one incident."
      },
...
      "RemediationPlaybook": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "risk_policy": {
            "type": "string",
            "enum": [
              "allow",
              "approval_required",
              "forbidden"
            ],
            "title": "Risk Policy",
            "default": "approval_required"
          },
          "approval_required": {
            "type": "boolean",
            "title": "Approval Required",
            "default": true
          },
          "pre_check": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pre Check"
          },
          "dry_run": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dry Run"
          },
          "sandbox_or_manual_record": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sandbox Or Manual Record"
          },
          "rollback": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rollback"
          },
          "observe_metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Observe Metrics"
          },
          "stop_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Stop Conditions"
          },
          "safety_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Safety Notes"
          }
        },
        "type": "object",
        "title": "RemediationPlaybook",
        "description": "Auditable remediation playbook that never executes production changes itself."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}